requirements.txt
google-generativeai
python-dotenv
PyPDF2
langchain
langchain-google-genai
faiss-cpu
streamlit
sentence-transformers